+++ src/invoke_training/training/shared/data/ARCHITECTURE.md
