+++ src/invoke_training/training/shared/data/base_image_caption_reader.py
