+++ src/invoke_training/training/shared/data/image_caption_sd_dataset.py
@@ -1,7 +1,7 @@
 from torchvision import transforms
 from transformers import CLIPTokenizer
 
-from invoke_training.training.shared.datasets.base_image_caption_reader import (
+from invoke_training.training.shared.data.base_image_caption_reader import (
     BaseImageCaptionReader,
 )
