+++ tests/invoke_training/training/shared/datasets/test_image_caption_sd_dataloader.py
@@ -5,7 +5,7 @@ import torch
 from transformers import CLIPTokenizer
 
 from invoke_training.training.finetune_lora.finetune_lora_config import DatasetConfig
-from invoke_training.training.shared.datasets.image_caption_sd_dataloader import (
+from invoke_training.training.shared.data.image_caption_sd_dataloader import (
     build_image_caption_sd_dataloader,
 )
